requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.103.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==0.21.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing-extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
uvloop==0.18.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.15.0
